requirements.txt
fastapi==0.115.6
pydantic==2.10.3
mysql-connector-python==9.1.0
jupyter_server==2.14.2
pandas==2.2.3
streamlit==1.40.2
requests==2.32.3
#requirements.txt file will be required if app is deployed in cloud using Heroku/ sharing with others
#pip install -r .\requirements.txt this will install all these modules